refactor(locale): rename pt-BR streetSuffix to streetPrefix
