+++ src/locales/pt_BR/location/index.ts
@@ -13,7 +13,7 @@ import secondary_address from './secondary_address';
 import state from './state';
 import state_abbr from './state_abbr';
 import street_pattern from './street_pattern';
-import street_suffix from './street_suffix';
+import street_prefix from './street_prefix';
 
 const location: LocationDefinition = {
   building_number,
@@ -26,7 +26,7 @@ const location: LocationDefinition = {
   state,
   state_abbr,
   street_pattern,
-  street_suffix,
+  street_prefix,
 };
 
 export default location;
