+++ src/locales/pt_BR/location/street_pattern.ts
@@ -1,4 +1,4 @@
 export default [
-  '{{person.firstName}} {{location.street_suffix}}',
-  '{{person.lastName}} {{location.street_suffix}}',
+  '{{location.street_prefix}} {{person.firstName}}',
+  '{{location.street_prefix}} {{person.lastName}}',
 ];
